+++ .gitmodules
@@ -1,6 +1,6 @@
-[submodule "Driver/CartPoleSimulation"]
-	path = Driver/CartPoleSimulation
-	url = https://github.com/SensorsINI/CartPoleSimulation
 [submodule "Driver/Control_Toolkit"]
 	path = Driver/Control_Toolkit
 	url = git@github.com:SensorsINI/Control_Toolkit.git
+[submodule "Driver/SI_Toolkit"]
+	path = Driver/SI_Toolkit
+	url = git@github.com:SensorsINI/SI_Toolkit.git
