+++ Driver/Control_Toolkit_ASF/Controllers/template_controller_from_cartpole_simulator.py
@@ -1,5 +1,4 @@
-from CartPoleSimulation.Controllers.controller_neural_imitator_tf import controller_neural_imitator_tf
-from Controllers.template_controller import template_controller
+from Driver.Control_Toolkit.Controllers import template_controller
 
 """
 Please modify this template to include a controller from cartpole simulator in physical cartpole
@@ -12,15 +11,17 @@ For a controller to be found and imported it must:
 4. It must have __init__ and step methods
 """
 
+class controller_from_cartpole_simulator_PCP(template_controller):
+    def __init__(self, SomeController, *args, **kwargs):
+        self.controller_kern: template_controller = SomeController(*args, **kwargs)
 
-class controller_neural_imitator_tf_CPS(controller_neural_imitator_tf):
-    def __init__(self):
-        super().__init__()
-        self.controller_name = "neural-imitator-tf-CPS"
+    def step(self, s, time=None):
+        Q = self.controller_kern.step(s, time)
+        return Q
+
+    def controller_reset(self):
+        self.controller_kern.controller_reset()
 
     def functions_specific_to_physical_cartpole(self):
         # e.g. inout and output from the terminal
-        pass
-
-    def controller_reset(self):
         pass
