+++ Driver/Control_Toolkit_ASF/models_for_nn_as_mpc_tf/Dense-6IN-128H1-128H2-1OUT-0/checkpoint
