+++ Driver/Control_Toolkit_ASF/models_for_nn_as_mpc_tf/Dense-6IN-32H1-32H2-1OUT-0/Dense-6IN-32H1-32H2-1OUT-0.txt
@@ -1,0 +1,41 @@
+CREATED:
+2021-11-01 at time 16:00:22
+
+WITH GIT REVISION:
+8e6eaf0c2305afc64a2289ac8c137883460eb3d6
+
+NET NAME:
+Dense-32H1-32H2
+
+NET FULL NAME:
+Dense-6IN-32H1-32H2-1OUT-0
+
+INPUTS:
+angleD, angle_cos, angle_sin, position, positionD, target_position
+
+OUTPUTS:
+Q
+
+TYPE:
+Dense
+
+NORMALIZATION:
+./Controllers/models_for_nn_as_mpc_tf/Dense-6IN-32H1-32H2-1OUT-0/NI_2021-11-01_13-53-51.csv
+
+SAMPLING INTERVAL:
+0.01 s
+
+PARENT NET:
+Network trained from scratch
+
+WASH OUT LENGTH:
+10
+
+TRAINING_FILES:
+./SI_Toolkit_ASF/Experiments/Imitation-lqr//Recordings/Train/
+
+VALIDATION_FILES:
+./SI_Toolkit_ASF/Experiments/Imitation-lqr//Recordings/Validate/
+
+TEST_FILES:
+./SI_Toolkit_ASF/Experiments/Imitation-lqr//Recordings/Test/
