+++ Driver/Control_Toolkit_ASF/models_for_nn_as_mpc_tf/Dense-6IN-32H1-32H2-1OUT-0/NI_2021-11-01_13-53-51.csv
@@ -1,0 +1,62 @@
+# This is normalization information calculated 2021-11-01 at time 13-53-51
+# Done with git-revision: d316affa6421014cd05e8596f665183cb27530b9
+#
+# Sampling interval of data used to calculate normalization: 0.01 s
+#
+# Data files used to calculate normalization information:
+#     ./SI_Toolkit_ASF/Experiments/Imitation-lqr/Recordings/Train/Experiment-1.csv
+#     ./SI_Toolkit_ASF/Experiments/Imitation-lqr/Recordings/Train/Experiment-10.csv
+#     ./SI_Toolkit_ASF/Experiments/Imitation-lqr/Recordings/Train/Experiment-11.csv
+#     ./SI_Toolkit_ASF/Experiments/Imitation-lqr/Recordings/Train/Experiment-12.csv
+#     ./SI_Toolkit_ASF/Experiments/Imitation-lqr/Recordings/Train/Experiment-13.csv
+#     ./SI_Toolkit_ASF/Experiments/Imitation-lqr/Recordings/Train/Experiment-14.csv
+#     ./SI_Toolkit_ASF/Experiments/Imitation-lqr/Recordings/Train/Experiment-15.csv
+#     ./SI_Toolkit_ASF/Experiments/Imitation-lqr/Recordings/Train/Experiment-16.csv
+#     ./SI_Toolkit_ASF/Experiments/Imitation-lqr/Recordings/Train/Experiment-17.csv
+#     ./SI_Toolkit_ASF/Experiments/Imitation-lqr/Recordings/Train/Experiment-18.csv
+#     ./SI_Toolkit_ASF/Experiments/Imitation-lqr/Recordings/Train/Experiment-19.csv
+#     ./SI_Toolkit_ASF/Experiments/Imitation-lqr/Recordings/Train/Experiment-2.csv
+#     ./SI_Toolkit_ASF/Experiments/Imitation-lqr/Recordings/Train/Experiment-20.csv
+#     ./SI_Toolkit_ASF/Experiments/Imitation-lqr/Recordings/Train/Experiment-21.csv
+#     ./SI_Toolkit_ASF/Experiments/Imitation-lqr/Recordings/Train/Experiment-22.csv
+#     ./SI_Toolkit_ASF/Experiments/Imitation-lqr/Recordings/Train/Experiment-23.csv
+#     ./SI_Toolkit_ASF/Experiments/Imitation-lqr/Recordings/Train/Experiment-24.csv
+#     ./SI_Toolkit_ASF/Experiments/Imitation-lqr/Recordings/Train/Experiment-25.csv
+#     ./SI_Toolkit_ASF/Experiments/Imitation-lqr/Recordings/Train/Experiment-26.csv
+#     ./SI_Toolkit_ASF/Experiments/Imitation-lqr/Recordings/Train/Experiment-27.csv
+#     ./SI_Toolkit_ASF/Experiments/Imitation-lqr/Recordings/Train/Experiment-28.csv
+#     ./SI_Toolkit_ASF/Experiments/Imitation-lqr/Recordings/Train/Experiment-29.csv
+#     ./SI_Toolkit_ASF/Experiments/Imitation-lqr/Recordings/Train/Experiment-3.csv
+#     ./SI_Toolkit_ASF/Experiments/Imitation-lqr/Recordings/Train/Experiment-30.csv
+#     ./SI_Toolkit_ASF/Experiments/Imitation-lqr/Recordings/Train/Experiment-31.csv
+#     ./SI_Toolkit_ASF/Experiments/Imitation-lqr/Recordings/Train/Experiment-32.csv
+#     ./SI_Toolkit_ASF/Experiments/Imitation-lqr/Recordings/Train/Experiment-33.csv
+#     ./SI_Toolkit_ASF/Experiments/Imitation-lqr/Recordings/Train/Experiment-34.csv
+#     ./SI_Toolkit_ASF/Experiments/Imitation-lqr/Recordings/Train/Experiment-35.csv
+#     ./SI_Toolkit_ASF/Experiments/Imitation-lqr/Recordings/Train/Experiment-36.csv
+#     ./SI_Toolkit_ASF/Experiments/Imitation-lqr/Recordings/Train/Experiment-37.csv
+#     ./SI_Toolkit_ASF/Experiments/Imitation-lqr/Recordings/Train/Experiment-38.csv
+#     ./SI_Toolkit_ASF/Experiments/Imitation-lqr/Recordings/Train/Experiment-39.csv
+#     ./SI_Toolkit_ASF/Experiments/Imitation-lqr/Recordings/Train/Experiment-4.csv
+#     ./SI_Toolkit_ASF/Experiments/Imitation-lqr/Recordings/Train/Experiment-5.csv
+#     ./SI_Toolkit_ASF/Experiments/Imitation-lqr/Recordings/Train/Experiment-6.csv
+#     ./SI_Toolkit_ASF/Experiments/Imitation-lqr/Recordings/Train/Experiment-7.csv
+#     ./SI_Toolkit_ASF/Experiments/Imitation-lqr/Recordings/Train/Experiment-8.csv
+#     ./SI_Toolkit_ASF/Experiments/Imitation-lqr/Recordings/Train/Experiment-9.csv
+#     ./SI_Toolkit_ASF/Experiments/Imitation-lqr/Recordings/Train/Experiment.csv
+#
+# Original (calculated from data) Normalization Information:
+#,         ,      ,         ,Q,         ,angle,         ,angleD,         ,angleDD,         ,angle_cos,         ,angle_sin,         ,position,         ,positionD,         ,positionDD,         ,target_position,         ,u,         
+#,           ,mean,           ,-2e-05,           ,0.0,           ,-0.0,           ,-0.0,           ,0.99992,           ,0.0,           ,3e-05,           ,-2e-05,           ,-2e-05,           ,-1e-05,           ,-0.00014,           
+#,           ,std,           ,0.0771,           ,0.0127,           ,0.05457,           ,0.97297,           ,0.00016,           ,0.01269,           ,0.05957,           ,0.07281,           ,0.49874,           ,0.10049,           ,0.47876,           
+#,           ,max,           ,0.46781,           ,0.05065,           ,0.34545,           ,15.25703,           ,1.0,           ,0.05063,           ,0.17738,           ,0.29676,           ,7.01274,           ,0.1584,           ,2.90509,           
+#,           ,min,           ,-0.46926,           ,-0.051,           ,-0.34542,           ,-15.25892,           ,0.9987,           ,-0.05098,           ,-0.17862,           ,-0.29753,           ,-7.00928,           ,-0.1584,           ,-2.91411,           
+#
+# Does user modified normalization info calculated from data?: Yes
+#
+# Normalization Information:
+,Q,angle,angleD,angleDD,angle_cos,angle_sin,position,positionD,positionDD,target_position,u
+mean,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0
+std,0.0771,0.0127,0.05457,0.97297,0.00016,0.01269,0.05957,0.07281,0.49874,0.10049,0.47876
+max,0.46926,0.051,0.34545,15.25892,1.0,0.05098,0.198,0.29753,7.01274,0.198,6.21
+min,-0.46926,-0.051,-0.34545,-15.25892,0.9987,-0.05098,-0.198,-0.29753,-7.01274,-0.198,-6.21
