+++ Driver/Control_Toolkit_ASF/models_for_nn_as_mpc_tf/Dense-6IN-32H1-32H2-1OUT-0/checkpoint
@@ -1,0 +1,2 @@
+model_checkpoint_path: "ckpt.ckpt"
+all_model_checkpoint_paths: "ckpt.ckpt"
