+++ Driver/Control_Toolkit_ASF/models_for_nn_as_mpc_tf/Dense-6IN-32H1-32H2-1OUT-0/log_training.csv
@@ -1,0 +1,6 @@
+epoch;loss;lr;val_loss
+0;0.001511988346464932;0.001;0.0008310140110552311
+1;0.0009324187994934618;0.001;0.0008291480480693281
+2;0.0009176880121231079;0.000100000005;0.0007971837185323238
+3;0.0009179054759442806;1e-04;0.0007977860514074564
+4;0.0009176989551633596;1e-04;0.0007983031682670116
