+++ Driver/config.yml
@@ -197,7 +197,6 @@ controller:
     PATH_TO_MODELS: './Controllers/models_for_nn_imitator_tf/'
 cartpole:
   SEED: 1873  # This is a seed for rng for CartPole instance class only. If "None" random seed based on datetime is used
-  PATH_TO_CONTROLLERS: './Controllers/'  # Path where controllers are stored
   PATH_TO_EXPERIMENT_RECORDINGS_DEFAULT: './CartPoleSimulation/Experiment_Recordings/'   # Where to save experiment recording per default
   m: 0.087  # mass of pole, kg # Checked by Antonio & Tobi
   M: 0.230  # mass of cart, kg # Checked by Antonio
